Separated matrix generation from utilities.
