+++ test/comps/test_qb.cc
@@ -1,6 +1,7 @@
 #include "RandLAPACK.hh"
 #include "rl_blaspp.hh"
 #include "rl_lapackpp.hh"
+#include "rl_gen.hh"
 
 #include <RandBLAS.hh>
 
@@ -245,10 +246,10 @@ TEST_F(TestQB, Polynomial_Decay_general1)
     QBTestData<double> all_data(m, n, k);
     algorithm_objects<double, r123::Philox4x32> all_algs(verbosity, cond_check, orth_check, p, passes_per_iteration, state);
 
-    RandLAPACK::util::mat_gen_info<double> m_info(m, n, RandLAPACK::util::polynomial);
+    RandLAPACK::gen::mat_gen_info<double> m_info(m, n, RandLAPACK::gen::polynomial);
     m_info.cond_num = 2025;
     m_info.rank = k;
-    RandLAPACK::util::mat_gen<double, r123::Philox4x32>(m_info, all_data.A, state);
+    RandLAPACK::gen::mat_gen<double, r123::Philox4x32>(m_info, all_data.A, state);
     
     svd_and_copy_computational_helper<double>(all_data);
     test_QB2_low_exact_rank<double, r123::Philox4x32>(block_sz, tol, all_data, all_algs);
@@ -273,10 +274,10 @@ TEST_F(TestQB, Polynomial_Decay_general2)
     QBTestData<double> all_data(m, n, k);
     algorithm_objects<double, r123::Philox4x32> all_algs(verbosity, cond_check, orth_check, p, passes_per_iteration, state);
     
-    RandLAPACK::util::mat_gen_info<double> m_info(m, n, RandLAPACK::util::polynomial);
+    RandLAPACK::gen::mat_gen_info<double> m_info(m, n, RandLAPACK::gen::polynomial);
     m_info.cond_num = 6.7;
     m_info.rank = k;
-    RandLAPACK::util::mat_gen<double, r123::Philox4x32>(m_info, all_data.A, state);
+    RandLAPACK::gen::mat_gen<double, r123::Philox4x32>(m_info, all_data.A, state);
 
     svd_and_copy_computational_helper<double>(all_data);
     test_QB2_low_exact_rank<double, r123::Philox4x32>(block_sz, tol, all_data, all_algs);
@@ -301,10 +302,10 @@ TEST_F(TestQB, Polynomial_Decay_zero_tol1)
     QBTestData<double> all_data(m, n, k);
     algorithm_objects<double, r123::Philox4x32> all_algs(verbosity, cond_check, orth_check, p, passes_per_iteration, state);
     
-    RandLAPACK::util::mat_gen_info<double> m_info(m, n, RandLAPACK::util::polynomial);
+    RandLAPACK::gen::mat_gen_info<double> m_info(m, n, RandLAPACK::gen::polynomial);
     m_info.cond_num = 2025;
     m_info.rank = k;
-    RandLAPACK::util::mat_gen<double, r123::Philox4x32>(m_info, all_data.A, state);
+    RandLAPACK::gen::mat_gen<double, r123::Philox4x32>(m_info, all_data.A, state);
 
     double norm_A = lapack::lange(Norm::Fro, m, n, all_data.A.data(), m);
     test_QB2_k_eq_min<double, r123::Philox4x32>(block_sz, tol, norm_A, all_data, all_algs);
@@ -329,10 +330,10 @@ TEST_F(TestQB, Polynomial_Decay_zero_tol2)
     QBTestData<double> all_data(m, n, k);
     algorithm_objects<double, r123::Philox4x32> all_algs(verbosity, cond_check, orth_check, p, passes_per_iteration, state);
     
-    RandLAPACK::util::mat_gen_info<double> m_info(m, n, RandLAPACK::util::polynomial);
+    RandLAPACK::gen::mat_gen_info<double> m_info(m, n, RandLAPACK::gen::polynomial);
     m_info.cond_num = 2025;
     m_info.rank = k;
-    RandLAPACK::util::mat_gen<double, r123::Philox4x32>(m_info, all_data.A, state);
+    RandLAPACK::gen::mat_gen<double, r123::Philox4x32>(m_info, all_data.A, state);
 
     double norm_A = lapack::lange(Norm::Fro, m, n, all_data.A.data(), m);
     test_QB2_k_eq_min<double, r123::Philox4x32>(block_sz, tol, norm_A, all_data, all_algs);
