+++ test/comps/test_util.cc
@@ -111,11 +111,11 @@ TEST_F(TestUtil, test_spectral_norm_polynomial_decay_double_precision) {
     auto state = RandBLAS::RNGState();
     SpectralTestData<double> all_data(m, n);
 
-    RandLAPACK::util::mat_gen_info<double> m_info(m, n, RandLAPACK::util::polynomial);
+    RandLAPACK::gen::mat_gen_info<double> m_info(m, n, RandLAPACK::gen::polynomial);
     m_info.cond_num = 2025;
     m_info.rank = n;
 
-    RandLAPACK::util::mat_gen<double, r123::Philox4x32>(m_info, all_data.A, state);
+    RandLAPACK::gen::mat_gen<double, r123::Philox4x32>(m_info, all_data.A, state);
     //lapack::lacpy(MatrixType::General, m, n, all_data.A.data(), m, all_data.A_cpy.data(), m);
     //test_spectral_norm<double, r123::Philox4x32>(state, all_data);
 }
@@ -127,11 +127,11 @@ TEST_F(TestUtil, test_spectral_norm_rank_def_mat_double_precision) {
     auto state = RandBLAS::RNGState();
     SpectralTestData<double> all_data(m, n);
 
-    RandLAPACK::util::mat_gen_info<double> m_info(m, n, RandLAPACK::util::adverserial);
+    RandLAPACK::gen::mat_gen_info<double> m_info(m, n, RandLAPACK::gen::adverserial);
     m_info.scaling = std::pow(10, 15);
     m_info.rank = n;
 
-    RandLAPACK::util::mat_gen<double, r123::Philox4x32>(m_info, all_data.A, state);
+    RandLAPACK::gen::mat_gen<double, r123::Philox4x32>(m_info, all_data.A, state);
     lapack::lacpy(MatrixType::General, m, n, all_data.A.data(), m, all_data.A_cpy.data(), m);
     test_spectral_norm<double, r123::Philox4x32>(state, all_data);
 }
@@ -143,11 +143,11 @@ TEST_F(TestUtil, test_spectral_norm_polynomial_decay_single_precision) {
     auto state = RandBLAS::RNGState();
     SpectralTestData<float> all_data(m, n);
 
-    RandLAPACK::util::mat_gen_info<float> m_info(m, n, RandLAPACK::util::polynomial);
+    RandLAPACK::gen::mat_gen_info<float> m_info(m, n, RandLAPACK::gen::polynomial);
     m_info.cond_num = 2;
     m_info.rank = n;
 
-    RandLAPACK::util::mat_gen<float, r123::Philox4x32>(m_info, all_data.A, state);
+    RandLAPACK::gen::mat_gen<float, r123::Philox4x32>(m_info, all_data.A, state);
     lapack::lacpy(MatrixType::General, m, n, all_data.A.data(), m, all_data.A_cpy.data(), m);
     test_spectral_norm<float, r123::Philox4x32>(state, all_data);
 }
@@ -159,11 +159,11 @@ TEST_F(TestUtil, test_spectral_norm_rank_def_mat_single_precision) {
     auto state = RandBLAS::RNGState();
     SpectralTestData<float> all_data(m, n);
 
-    RandLAPACK::util::mat_gen_info<float> m_info(m, n, RandLAPACK::util::adverserial);
+    RandLAPACK::gen::mat_gen_info<float> m_info(m, n, RandLAPACK::gen::adverserial);
     m_info.scaling = std::pow(10, 7);
     m_info.rank = n;
 
-    RandLAPACK::util::mat_gen<float, r123::Philox4x32>(m_info, all_data.A, state);
+    RandLAPACK::gen::mat_gen<float, r123::Philox4x32>(m_info, all_data.A, state);
     lapack::lacpy(MatrixType::General, m, n, all_data.A.data(), m, all_data.A_cpy.data(), m);
     test_spectral_norm<float, r123::Philox4x32>(state, all_data);
 }
