+++ test/drivers/test_cqrrpt.cc
@@ -1,6 +1,7 @@
 #include "RandLAPACK.hh"
 #include "rl_blaspp.hh"
 #include "rl_lapackpp.hh"
+#include "rl_gen.hh"
 
 #include <RandBLAS.hh>
 #include <fstream>
@@ -141,10 +142,10 @@ TEST_F(TestCQRRPT, CQRRPT_full_rank_no_hqrrp) {
     CQRRPT.num_threads = 4;
     CQRRPT.no_hqrrp = 1;
 
-    RandLAPACK::util::mat_gen_info<double> m_info(m, n, RandLAPACK::util::polynomial);
+    RandLAPACK::gen::mat_gen_info<double> m_info(m, n, RandLAPACK::gen::polynomial);
     m_info.cond_num = 2;
     m_info.rank = k;
-    RandLAPACK::util::mat_gen<double, r123::Philox4x32>(m_info, all_data.A, state);
+    RandLAPACK::gen::mat_gen<double, r123::Philox4x32>(m_info, all_data.A, state);
 
     norm_and_copy_computational_helper<double, r123::Philox4x32>(norm_A, all_data);
     test_CQRRPT_general<double, r123::Philox4x32>(d, norm_A, all_data, CQRRPT);
@@ -165,10 +166,10 @@ TEST_F(TestCQRRPT, CQRRPT_low_rank_with_hqrrp) {
     CQRRPT.num_threads = 4;
     CQRRPT.no_hqrrp = 0;
 
-    RandLAPACK::util::mat_gen_info<double> m_info(m, n, RandLAPACK::util::polynomial);
+    RandLAPACK::gen::mat_gen_info<double> m_info(m, n, RandLAPACK::gen::polynomial);
     m_info.cond_num = 2;
     m_info.rank = k;
-    RandLAPACK::util::mat_gen<double, r123::Philox4x32>(m_info, all_data.A, state);
+    RandLAPACK::gen::mat_gen<double, r123::Philox4x32>(m_info, all_data.A, state);
 
     norm_and_copy_computational_helper<double, r123::Philox4x32>(norm_A, all_data);
     test_CQRRPT_general<double, r123::Philox4x32>(d, norm_A, all_data, CQRRPT);
@@ -191,9 +192,9 @@ TEST_F(TestCQRRPT, CQRRPT_bad_orth) {
     CQRRPT.num_threads = 4;
     CQRRPT.no_hqrrp = 1;
 
-    RandLAPACK::util::mat_gen_info<double> m_info(m, n, RandLAPACK::util::adverserial);
+    RandLAPACK::gen::mat_gen_info<double> m_info(m, n, RandLAPACK::gen::adverserial);
     m_info.scaling = 1e7;
-    RandLAPACK::util::mat_gen<double, r123::Philox4x32>(m_info, all_data.A, state);
+    RandLAPACK::gen::mat_gen<double, r123::Philox4x32>(m_info, all_data.A, state);
 
     norm_and_copy_computational_helper<double, r123::Philox4x32>(norm_A, all_data);
     test_CQRRPT_general<double, r123::Philox4x32>(d, norm_A, all_data, CQRRPT);
