+++ test/drivers/test_revd2.cc
@@ -1,6 +1,7 @@
 #include "RandLAPACK.hh"
 #include "rl_blaspp.hh"
 #include "rl_lapackpp.hh"
+#include "rl_gen.hh"
 
 #include <RandBLAS.hh>
 #include <fstream>
@@ -244,10 +245,10 @@ TEST_F(TestREVD2, Underestimation1) {
         num_steps_power_iter_error_est
     );
 
-    RandLAPACK::util::mat_gen_info<double> m_info(m, m, RandLAPACK::util::polynomial);
+    RandLAPACK::gen::mat_gen_info<double> m_info(m, m, RandLAPACK::gen::polynomial);
     m_info.cond_num = std::pow(10, 8);
     m_info.rank = k;
-    RandLAPACK::util::mat_gen<double, r123::Philox4x32>(m_info, all_data.A_cpy, state);
+    RandLAPACK::gen::mat_gen<double, r123::Philox4x32>(m_info, all_data.A_cpy, state);
 
     symm_mat_and_copy_computational_helper<double, RNG>(norm_A, all_data);
     test_REVD2_general<double, RNG>(
@@ -281,10 +282,10 @@ TEST_F(TestREVD2, Underestimation2) {
         num_steps_power_iter_error_est
     );
 
-    RandLAPACK::util::mat_gen_info<double> m_info(m, m, RandLAPACK::util::polynomial);
+    RandLAPACK::gen::mat_gen_info<double> m_info(m, m, RandLAPACK::gen::polynomial);
     m_info.cond_num = std::pow(10, 8);
     m_info.rank = k;
-    RandLAPACK::util::mat_gen<double, r123::Philox4x32>(m_info, all_data.A_cpy, state);
+    RandLAPACK::gen::mat_gen<double, r123::Philox4x32>(m_info, all_data.A_cpy, state);
 
     symm_mat_and_copy_computational_helper<double, RNG>(norm_A, all_data);
     test_REVD2_general<double, RNG>(
@@ -318,10 +319,10 @@ TEST_F(TestREVD2, Overestimation1) {
         num_steps_power_iter_error_est
     );
 
-    RandLAPACK::util::mat_gen_info<double> m_info(m, m, RandLAPACK::util::polynomial);
+    RandLAPACK::gen::mat_gen_info<double> m_info(m, m, RandLAPACK::gen::polynomial);
     m_info.cond_num = std::pow(10, 2);
     m_info.rank = k;
-    RandLAPACK::util::mat_gen<double, r123::Philox4x32>(m_info, all_data.A_cpy, state);
+    RandLAPACK::gen::mat_gen<double, r123::Philox4x32>(m_info, all_data.A_cpy, state);
 
     symm_mat_and_copy_computational_helper<double, RNG>(norm_A, all_data);
     test_REVD2_general<double, RNG>(
@@ -355,10 +356,10 @@ TEST_F(TestREVD2, Oversetimation2) {
         num_steps_power_iter_error_est
     );
 
-    RandLAPACK::util::mat_gen_info<double> m_info(m, m, RandLAPACK::util::polynomial);
+    RandLAPACK::gen::mat_gen_info<double> m_info(m, m, RandLAPACK::gen::polynomial);
     m_info.cond_num = std::pow(10, 2);
     m_info.rank = k;
-    RandLAPACK::util::mat_gen<double, r123::Philox4x32>(m_info, all_data.A_cpy, state);
+    RandLAPACK::gen::mat_gen<double, r123::Philox4x32>(m_info, all_data.A_cpy, state);
 
     symm_mat_and_copy_computational_helper<double, RNG>(norm_A, all_data);
     test_REVD2_general<double, RNG>(
@@ -392,10 +393,10 @@ TEST_F(TestREVD2, Exactness) {
         num_steps_power_iter_error_est
     );
 
-    RandLAPACK::util::mat_gen_info<double> m_info(m, m, RandLAPACK::util::polynomial);
+    RandLAPACK::gen::mat_gen_info<double> m_info(m, m, RandLAPACK::gen::polynomial);
     m_info.cond_num = std::pow(10, 2);
     m_info.rank = k;
-    RandLAPACK::util::mat_gen<double, r123::Philox4x32>(m_info, all_data.A_cpy, state);
+    RandLAPACK::gen::mat_gen<double, r123::Philox4x32>(m_info, all_data.A_cpy, state);
 
     symm_mat_and_copy_computational_helper<double, RNG>(norm_A, all_data);
     test_REVD2_general<double, RNG>(
@@ -427,10 +428,10 @@ TEST_F(TestREVD2, Uplo) {
         num_steps_power_iter_error_est
     );
 
-    RandLAPACK::util::mat_gen_info<double> m_info(m, m, RandLAPACK::util::polynomial);
+    RandLAPACK::gen::mat_gen_info<double> m_info(m, m, RandLAPACK::gen::polynomial);
     m_info.cond_num = std::pow(10, 2);
     m_info.rank = k;
-    RandLAPACK::util::mat_gen<double, r123::Philox4x32>(m_info, all_data.work, state);
+    RandLAPACK::gen::mat_gen<double, r123::Philox4x32>(m_info, all_data.work, state);
 
     uplo_computational_helper<double, RNG>(all_data);
